Decompose underlying auth code into separate modules
